+++ src/pickDeliverE/src/pickDeliverEuclidean.c
@@ -36,7 +36,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_common/orders_input.h"
 #include "c_common/vehicles_input.h"
 
-#include "./pickDeliverEuclidean_driver.h"
+#include "drivers/pickDeliver/pickDeliverEuclidean_driver.h"
 
 PGDLLEXPORT Datum
 pickDeliverEuclidean(PG_FUNCTION_ARGS);
